Change uint to size_t

+++ InputBuffer.cc
@@ -1,11 +1,11 @@
 #include "InputBuffer.hh"
 
 
-InputBuffer::InputBuffer(uint max_capacity) : max_capacity(max_capacity), pallets_(max_capacity) {}
+InputBuffer::InputBuffer(size_t max_capacity) : max_capacity(max_capacity), pallets_(max_capacity) {}
 
 void
-InputBuffer::generateNewPallet(uint id, Height height) noexcept {
-  for(uint idx=0; idx<max_capacity; ++idx)
+InputBuffer::generateNewPallet(size_t id, Height height) noexcept {
+  for(size_t idx=0; idx<max_capacity; ++idx)
     if(!pallets_[idx]) {
       pallets_[idx]=std::make_unique<Pallet>(id,height);
       return;
@@ -13,17 +13,17 @@ InputBuffer::generateNewPallet(uint id, Height height) noexcept {
 }
 
 bool
-InputBuffer::isValid (uint position) const noexcept {
+InputBuffer::isValid (size_t position) const noexcept {
   return position < max_capacity;
 }
 
 const Pallet::Ptr&
-InputBuffer::getPallet (uint position) const noexcept{
+InputBuffer::getPallet (size_t position) const noexcept{
   return pallets_[position];
 }    
 
 bool
-InputBuffer::isApproachable(Height object_height, uint position) const noexcept(false){
+InputBuffer::isApproachable(Height object_height, size_t position) const noexcept(false){
   if (!isValid(position)) throw WrongPositionException();
   else return true;
 }
